migrated functions in sw into indexedDB file. Added logic to check for online /offline status
