+++ disco-db-demo/components/Notes.js
@@ -6,14 +6,9 @@ import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import CssBaseline from '@mui/material/CssBaseline';
 import { useRouter } from 'next/router';
-import { dexieUpdate, dexieDeleteOne } from '../public/db'
 
 export default function NotesContainer(props) {
-  // const [setContent, setNewContent] = useState('');
-  // const [setTitle, setNewTitle] = useState('');
-  //grab query paramaters from note - work in progress
-  // const router = useRouter();
-  // let noteID = Object.keys(router.query)[0];
+
 
   console.log('current setTitle: ',props.setTitle)
 
@@ -82,7 +77,6 @@ export default function NotesContainer(props) {
         content: noteContent.value,
         updatedAt: Date.now()
       }
-      //dexieUpdate(props.noteID, testData);
 
       props.setRefresh(true);
     })
@@ -121,9 +115,6 @@ export default function NotesContainer(props) {
     .then(data => {
       console.log('Success', data);
       //should remove entry from array
-
-      //testing - delete item from indexedDB
-      //dexieDeleteOne(props.noteID);
 
       props.setRefresh(true);
     })
