+++ disco-db-demo/components/Sidebar.js
@@ -11,7 +11,6 @@ import ListItem from '@mui/material/ListItem';
 import ListItemText from '@mui/material/ListItemText';
 import NotesIcon from '@mui/icons-material/Notes';
 import { Button } from '@mui/material';
-import { dexieAdd } from '../public/db'
 
 const drawerWidth = 240;
 
@@ -64,19 +63,6 @@ console.log('this is in sidebar.js', props.noteArray);
       console.log('this is the response from NEW NOTE button', data);
       const uniqId = data.data._id;
 
-      //testing - add new note to indexedDB
-      const testData = {
-        username: ['erictest02'],
-        _id: uniqId,
-        title: '',
-        content: '',
-        createdAt: Date.now(),
-        updatedAt: Date.now(),
-      }
-
-      //dexieAdd(testData);
-      //random num for testing purposes
-      //const ranNum = Math.ceil(Math.random() * 10)
       const newNote = 
           <ListItem button id={uniqId} key={uniqId} onClick={currNoteHandler}>
             <NotesIcon></NotesIcon>
