+++ disco-db-demo/components/layout.js
@@ -3,7 +3,6 @@ import Sidebar from './Sidebar';
 import Box from '@mui/material/Box';
 import CssBaseline from '@mui/material/CssBaseline';
 import Toolbar from '@mui/material/Toolbar';
-import { dexieAdd, dexieQuery, dexieDeleteAll } from '../public/db'
 import React, { useEffect, useState } from 'react';
 
 export default function Layout({ children }) {
@@ -24,7 +23,10 @@ export default function Layout({ children }) {
     return child;
   });
 
+
+
   useEffect(() => {
+
     const testURL = '/api/hello';
     const devURL = '/user/load';
     fetch(devURL)
@@ -32,37 +34,21 @@ export default function Layout({ children }) {
     .then( (data) => {
       //Iterate thru retrived data and create a copy of each object into state array.
 
-      // delete all user related data in indexDB using Dexie
-      // need to revisit later to see if delete query works with multiple users in username array
-      //dexieDeleteAll([localStorage.getItem('user')]);
-
       data.data.forEach((ele) => {
 
         userNoteArr.push(ele);
-        // add data to indexedDB using Dexie
-        const { username, _id, title, content, createdAt, updatedAt } = ele
-        //dexieAdd(ele);
+
+        //const { username, _id, title, content, createdAt, updatedAt } = ele
+
       });
         setNewNote(userNoteArr);
 
         setLoading(false);
         setRefresh(false);
 
-        //test to add data into indexedDB
-        // const ranNum = Math.random() * 100000
-        // const testData = {
-        //   username: ['erictest02'],
-        //   _id: ranNum,
-        //   title: 'testDexieAdd',
-        //   content: 'testDexieAddContent',
-        //   createdAt: Date.now(),
-        //   updatedAt: Date.now(),
-        // }
-        // dexieAdd(testData);
-        //dexieQuery([localStorage.getItem('user')]);
     })
     .catch((err) => console.log('Error in fetching data', err))
-  }, [refresh]);
+  }, [refresh, online]);
 
   if (isLoading) return null;
   else {
