+++ disco-db-demo/public/indexedDB.js
@@ -1,9 +1,10 @@
 
 const dbGlobals = {
   DB: null,
-  version: 1,
+  version: 8,
   databaseName: 'notesDB',
   storeName: 'notesStore',
+  failed_requests: 'failed_requests',
   keyPath: '_id'
 }
 
@@ -24,6 +25,12 @@ function openDB (callback) {
       db.createObjectStore(dbGlobals.storeName, {
         keyPath: dbGlobals.keyPath,
       });
+    }
+    if (!db.objectStoreNames.contains(dbGlobals.failed_requests)) {
+      console.log('Creating failed_request store')
+      db.createObjectStore(dbGlobals.failed_requests, {
+        keyPath: 'id', autoIncrement: true,
+      })
     }
   };
   req.onsuccess = (event) => {
@@ -74,7 +81,66 @@ function dbDeleteAll() {
   } else {
     console.log('DB is closed');
   }
+};
+
+//Function to Access specific object store in IDB database and start a transaction
+function accessObjectStore (storeName, method) {
+  return dbGlobals.DB.transaction([storeName], method).objectStore(storeName)
+};
+
+function patchData (data) {
+  //Open a transaction into store 'failed-requests' 
+  //Saves the persisting data in payload key
+  //URL and method to match type of request
+  const store = accessObjectStore(dbGlobals.failed_requests, 'readwrite')
+  store.add({url: '/user/notes', payload: data, method: 'PATCH'})
 }
 
+function deleteData (data) {
+  //Open a transaction into store 'failed-requests' 
+  const store = accessObjectStore(dbGlobals.failed_requests, 'readwrite')
+  store.add({url: '/user/notes', payload: data, method: 'DELETE'})
+}
 
-export { openDB, dbAdd, dbDeleteAll, dbGlobals }
+function postData (data) {
+  //Open a transaction into store 'failed-requests' 
+  const store = accessObjectStore(dbGlobals.failed_requests, 'readwrite')
+  store.add({url: '/user/notes', payload: data, method: 'POST'})
+}
+
+//Function to sync offline requests to database when client is back online
+function syncDataToServer() {
+  //Create transaction to object store and grab all objects in an ordered array by ID.
+  const store = accessObjectStore(dbGlobals.failed_requests, 'readwrite');
+  const request = store.getAll();
+
+  request.onsuccess = async function (event) {
+    const failedRequests = event.target.result;
+    //Comes back as an array of objects 
+    //Iterate through saved failed HTTP requests and creates a format to recreate a Fetch Request.
+    failedRequests.forEach((data) => {
+      const url = data.url;
+      const method = data.method;
+      const body = JSON.stringify(data.payload)
+      const headers = {'Content-Type': 'application/json'};
+      fetch(url, {
+        method: method,
+        headers: headers,
+        body: body
+      })
+      .then((res) => res.json())
+      .then((res) => {
+        //Previous transaction was closed due to getAll()
+        //Reopen object store and delete the corresponding object on successful HTTP request
+        const newStore = accessObjectStore(dbGlobals.failed_requests, 'readwrite');
+        newStore.delete(data.id);
+      })
+      .catch((error) => {
+        console.error('Failed to sync data to server:', error);
+        throw error
+      })
+    });
+  }
+};
+
+export { openDB, dbAdd, dbDeleteAll, patchData, deleteData, postData, syncDataToServer, dbGlobals }
